pyproject.toml
[project]
name = "nonebot-plugin-echo"
version = "0.1.0"
description = "NoneBot2 插件，让 Bot 学习群友的发言和表情包。/ NoneBot2 plugin to learn what your group members say."
authors = [
    { name = "KomoriDev", email = "mute231010@gmail.com" },
    { name = "CMHopeSunshine", email = "277073121@qq.com" },
]
license = { text = "AGPL-3" }
readme = "README.md"
keywords = []

requires-python = ">=3.9"
dependencies = [
    "nonebot-plugin-alconna>=0.59.0",
    "nonebot-plugin-apscheduler>=0.5.0",
    "nonebot-plugin-localstore>=0.7.4",
    "nonebot-plugin-orm>=0.8.2",
    "nonebot2>=2.4.2",
    "pydantic-yaml>=1.4.0",
]

[project.urls]
homepage = "https://github.com/KomoriDev/nonebot-plugin-echo"
repository = "https://github.com/KomoriDev/nonebot-plugin-echo"

[dependency-groups]
dev = [
    "isort>=6.0.1",
    "nonebot-adapter-onebot>=2.4.6",
    "nonebot-plugin-orm[aiosqlite]>=0.8.2",
    "nonebot2[fastapi]>=2.4.2",
    "pre-commit>=4.2.0",
    "ruff>=0.12.5",
]

[tool.nonebot]
adapters = [
    { name = "OneBot V11", module_name = "nonebot.adapters.onebot.v11" }
]
plugins = ["nonebot_plugin_echo"]

[tool.isort]
profile = "black"
line_length = 88
length_sort = true
skip_gitignore = true
force_sort_within_sections = true
extra_standard_library = ["typing_extensions"]

[tool.ruff]
line-length = 88
target-version = "py39"

[tool.ruff.format]
line-ending = "lf"

[tool.ruff.lint]
select = ["E", "W", "F", "UP", "C", "T", "PYI", "PT", "Q"]
ignore = ["E402", "C901"]

[tool.ruff.lint.isort]
length-sort = true
force-sort-within-sections = true
extra-standard-library = ["typing_extensions"]

[tool.ruff.lint.flake8-pytest-style]
fixture-parentheses = false
mark-parentheses = false

[tool.ruff.lint.pyupgrade]
keep-runtime-typing = true

[tool.pyright]
pythonVersion = "3.9"
pythonPlatform = "All"
typeCheckingMode = "standard"

[tool.uv.build-backend]
module-name = "nonebot_plugin_echo"
module-root = ""

[build-system]
requires = ["uv_build>=0.8.2,<0.9.0"]
build-backend = "uv_build"
